pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "email-verify"
version = "0.2.0"
description = "Email intelligence & SMTP verification CLI — DNS, Mailgun validation, GeoIP, ASN, risk scoring"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Craig Derington", email = "craig@gb10.studio" }
]
keywords = ["email", "validation", "smtp", "mailgun", "dns", "cli", "tui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: Utilities",
]
dependencies = [
    "dnspython>=2.6.0",
    "textual>=0.80.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "respx>=0.21.0",        # httpx mock library
    "ruff>=0.4.0",
    "pyright>=1.1.0",
]

[project.scripts]
email-verify = "email_verify.cli:main"

[project.urls]
Homepage    = "https://gb10.studio"
Repository  = "https://github.com/craigderington/email-verify"
Issues      = "https://github.com/craigderington/email-verify/issues"
Changelog   = "https://github.com/craigderington/email-verify/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/email_verify"]

[tool.hatch.build.targets.sdist]
include = ["src/", "README.md", "LICENSE", "pyproject.toml", "CHANGELOG.md"]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PIE"]
ignore = ["E501"]   # line length handled by formatter

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode   = "auto"
testpaths      = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

# ── Pyright ───────────────────────────────────────────────────────────────────
[tool.pyright]
include               = ["src"]
pythonVersion         = "3.11"
typeCheckingMode      = "basic"
reportMissingImports  = true
